requirements.txt
streamlit>=1.28.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
langchain-core>=0.2.0
